% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmcForwardBackward}
\alias{ctmcForwardBackward}
\title{Implementation of ForwardBackward Algorithm on a Symmetric Continuous-time Markov Chain HMM}
\usage{
ctmcForwardBackward(ts, theta, obs)
}
\arguments{
\item{ts}{size m Numeric vector containing the timestamp of each observation}

\item{theta}{double type value which represents the parameter theta in the model of transition probability}

\item{obs}{n x m Numeric matrix containing the likelihood of the observed data}
}
\value{
a n x m matrix containing the conditional probabilities of each observation
}
\description{
Implementing the ForwardBackward algorithm,given likelihood of observed data,timestamp of each observation and a parameter theta.
}
\examples{
obs <- matrix(c(0.88,0.10,0.88,0.10,0.02,0.30,0.02,0.30,0.10,0.60),2,5)
theta <- log(2)
ts <- c(1,2,3,4,5)
ctmcForwardBackward(ts,theta,obs)
}
