% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ctmcViterbi}
\alias{ctmcViterbi}
\title{Implementation of Viterbi Algorithm on a Symmetric Continuous-time Markov Chain HMM using Rcpp}
\usage{
ctmcViterbi(ts, theta, obs)
}
\arguments{
\item{ts}{size m Numeric vector containing the timestamp of each observation}

\item{theta}{double type value which represents the parameter theta in the model of transition probability}

\item{obs}{n x m Numeric matrix containing the likelihood of the observed data}
}
\value{
the Viterbi path(path)
}
\description{
Function that implements a Viterbi algorithm, given likelihood of observed data,timestamp of each observation and a parameter theta.
}
\examples{
obs <- matrix(c(0.88,0.10,0.88,0.10,0.02,0.30,0.02,0.30,0.10,0.60),2,5)
theta <- log(2)
ts <- c(1,2,3,4,5)
ctmcViterbi(ts,theta,obs)
}
