\name{gspyrouHW4-package}
\alias{gspyrouHW4-package}
\alias{gspyrouHW4}
\docType{package}
\title{
  Viterbi and ForwardBackward algorithms for Symmetric Continuous-time Markov-Chain HMM
}
\description{
  The package contains two main algorithms for generic Continuous-time Markov Chain hidden Markov Models
}
\details{
  Package: gspyrouHW4
  
  Type: Package
  
  Title: Viterbi and ForwardBackward algorithms for Symmetric Continuous-time Markov-Chain HMM
  
  Version: 1.0
  
  Date: 2017-11-19
  
  Author: Georgios Spyrou(gspyrou)
  
  Maintainer: Georgios Spyrou(gspyrou) <gspyrou@umich.edu>
  
  Description: This package contains two essential functions for Symmetric Continuous-time M-C HMM
  
  License: GPL (>= 3)
  
  Index of help topics:
  
  ctmcViterbi : Viterbi algorithm for Symmetric Continuous-time Markov-Chain HMM
  
  ctmcForwardBackward :  ForwardBackward algorithm for Symmetric Continuous-time Markov-Chain HMM


}
\author{
  Georgios Spyrou(gspyrou)
  
  Maintainer: Georgios Spyrou(gspyrou) <gspyrou@umich.edu>
}
\references{
  https://en.wikipedia.org/wiki/Viterbi_algorithm
  
  https://en.wikipedia.org/wiki/Forward-backward_algorithm

}
